% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_probability_dist.R
\name{compute_probability_dist}
\alias{compute_probability_dist}
\title{Compute probability distribution of a lengthy text}
\usage{
compute_probability_dist(text)
}
\arguments{
\item{text}{A lengthy string of text}
}
\value{
a data frame of the unique words, their frequency, and their probability (n/N)
}
\description{
Takes a string of text, returns a data frame of the unique words, their frequency, and their probability (n/N)
}
