% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentence_weights.R
\name{sentence_weights}
\alias{sentence_weights}
\title{Compute sentence weights}
\usage{
sentence_weights(text, dist)
}
\arguments{
\item{text}{A lengthy string of text}

\item{dist}{A probability distribution, generated by compute_probability_dist()}
}
\value{
A data frame, consisting of the sentences and their weights
}
\description{
For a given text and probability distribution, calculates the weights for each sentence
}
