% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumbasic.R
\name{sumbasic}
\alias{sumbasic}
\title{SumBasic: A Frequency-Based Summarization System}
\usage{
sumbasic(text, len = 3)
}
\arguments{
\item{text}{A lengthy string of text}

\item{len}{Length (in sentences) of the desired summary}
}
\value{
The summary, generated via SumBasic
}
\description{
Takes a string of text and a length.  Returns a summary of the text of the given length via SumBasic.
}
